requirements.txt
fastapi
uvicorn
Pillow
torch
torchvision
pymilvus
scikit-learn
python-multipart
timm